@@ -1,9 +1,13 @@
 name: effective_dart_rules
-version: 0.0.2
+version: 0.0.3
 description: Builds a file with all the rules in the different sections of Effective Dart.
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
+
 dependencies:
-  html_unescape: ^1.0.0
-  markdown: ^2.0.0
-  path: ^1.6.0
+  html_unescape: ^2.0.0
+  markdown: ^4.0.0
+  path: ^1.8.0
+
+dev_dependencies:
+  lints: ^1.0.0
